add Dockerfile & docker-compose.yml
The Dockerfile should setup a fully working environment for kippo,
including optional dependencies (mysql, xmpp).

To build:

  docker build -t kippo:latest .

To use:

  docker run \
      -it \
      --rm \
      --name=kippo \
      -p 2222:2222 \
      -v $(pwd)/log:/app/log \
      -v $(pwd)/dl:/app/dl \
      -v $(pwd)/data:/app/data \
      -v $(pwd)/kippo.cfg:/app/kippo.cfg \
      kippo

With the docker-compose.yml, kippo can be run with a single command:

  docker-compose up

To start with the optional mariadb service:

  docker-compose --profile with-db up

All data and configuration should be kept outside the container by using
volumes. The docker-compose.yml has these volumes pointing to the usual
directories in the main kippo directory.